fix up organizations page
