+++ src/containers/Grants.js
@@ -2,6 +2,30 @@ import React from 'react';
 
 import { Switch, Route, useRouteMatch, useParams } from 'react-router-dom';
 
+import { gql } from 'apollo-boost';
+import { useQuery } from '@apollo/react-hooks';
+
+const GET_GRANT = gql`
+  query grant($grantId: String!) {
+    grant(uuid: $grantId) {
+      uuid
+      dateFrom
+      dateTo
+      amount
+      source
+      description
+      from {
+        name
+        uuid
+      }
+      to {
+        name
+        uuid
+      }
+    }
+  }
+`;
+
 export default () => {
   let match = useRouteMatch();
 
@@ -22,5 +46,18 @@ export default () => {
 
 const Grant = () => {
   let { grantId } = useParams();
-  return <h3>Requested grant ID: {grantId}</h3>;
+
+  const { loading, error, data } = useQuery(GET_GRANT, {
+    variables: { grantId },
+  });
+
+  if (loading) return 'Loading...';
+  if (error) return `Error! ${error}`;
+
+  return (
+    <div>
+      <h3>Requested grant ID: {grantId}</h3>
+      <pre>{JSON.stringify(data)}</pre>
+    </div>
+  );
 };
